Updates for API changes
The API is now reporting the latest recorded EGV on the dataRange
endpoint, but still has a 3-hour lag on the egvs endpoint, which means
we weren't getting any data. This change makes it so we assume that
latest_egv is 4 hours ahead so we make sure we don't miss anything.